% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wl8_diagram.R
\name{wl8_diagram}
\alias{wl8_diagram}
\title{Generate Win Loss Diagram}
\usage{
wl8_diagram(data, topline, colour = TRUE)
}
\arguments{
\item{data}{Data frame with columns `endpoint`, ` 'wins', 'ties', 'losses'`.
The dataframe should have exactly 8 rows.}

\item{topline}{Character string to be included in the top box.}

\item{colour}{Logical default = TRUE. If TRUE the diagram will be coloured.}
}
\value{
grViz object. Refer to DiagrammeR package for details.
}
\description{
Uses the DiagrammeR package to return a win-loss diagram.
Currently only supports exactly 8 rows.
}
\examples{
data <- data.frame(endpoint = c(paste0("endpoint", 1:8)),
wins = c(5,5,5,5,5,5,5,5),
ties = c(80,70,60,50,40,30,20,10),
losses = c(5,5,5,5,5,5,5,5))
wl8_diagram(data, topline="N=10*9 = 90")
wl8_diagram(data, topline="N=10*9 = 90", colour = FALSE)
}
