% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wl_diagram.R
\name{wl_diagram}
\alias{wl_diagram}
\title{Generate Win Loss Diagram}
\usage{
wl_diagram(data, topline)
}
\arguments{
\item{data}{Data frame with columns `endpoint`, ` 'wins', 'ties', 'losses'`.
The dataframe should have exactly 5 rows.}

\item{topline}{Character string to be included in the top box.}
}
\value{
grViz object. Refer to DiagrammeR package for details.
}
\description{
Uses the DiagrammeR package to return a win-loss diagram.
Currently only supports exactly 5 rows.
}
\examples{
data <- data.frame(endpoint = c(paste0("endpoint", 1:5)),
wins = c(5,5,5,5,5),
ties = c(50,40,30,20,10),
losses = c(5,5,5,5,5))
wl_diagram(data, topline="N=10*6 = 60")

}
